backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
asyncpg==0.29.0
psycopg[binary]==3.1.19
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
pgvector==0.2.5
langchain==0.2.3
langchain-postgres==0.0.6
langchain-community==0.2.4
openai==1.30.1
httpx==0.27.0
dashscope==1.20.14
pdfplumber==0.11.0
python-docx==1.1.2
structlog==24.2.0
